2019/5/21

+++ src/main/java/com/dt/project/service/SystemService/SystemUserConfigService.java
@@ -17,7 +17,7 @@ public interface SystemUserConfigService {
     /**
      * 新增一个用户配置信息
      */
-    Long serviceSaveUserConfig(SystemUserConfig userConfig);
+    Integer serviceSaveUserConfig(SystemUserConfig userConfig);
 
     /**
      * 新增用户配置信息
